refactor(access-key): split transaction lifecycle
