+++ src/cli/adapter.ts
@@ -14,8 +14,9 @@ import { KeyAuthorization } from 'ox/tempo'
 import { Account as TempoAccount, Secp256k1 } from 'viem/tempo'
 import * as z from 'zod/mini'
 
-import * as AccessKey from '../core/AccessKey.js'
 import * as Adapter from '../core/Adapter.js'
+import * as AccessKeyStore from '../core/internal/AccessKeyStore.js'
+import * as AccessKeyTransaction from '../core/internal/AccessKeyTransaction.js'
 import * as CliAuth from '../server/CliAuth.js'
 import * as Keyring from './keyring.js'
 
@@ -44,21 +45,13 @@ export function cli(options: cli.Options): Adapter.Adapter {
       if (!deserialized.signature) throw new Error('Managed access key is missing a signature.')
       const keyAuthorization = deserialized as KeyAuthorization.Signed
       const keyAuthorizationStatus = entry.keyAuthorizationStatus ?? 'pending'
-      if (keyAuthorizationStatus === 'authorized')
-        AccessKey.saveAuthorized({
-          address,
-          keyAuthorization,
-          privateKey: entry.key,
-          store,
-        })
-      else
-        AccessKey.savePending({
-          address,
-          keyAuthorization,
-          ...(keyAuthorizationStatus === 'pending' ? { keyAuthorizationPending: true } : {}),
-          privateKey: entry.key,
-          store,
-        })
+      AccessKeyStore.upsertAuthorization({
+        address,
+        keyAuthorization,
+        privateKey: entry.key,
+        state: keyAuthorizationStatus,
+        store,
+      })
 
       return { ...entry, keyAuthorizationStatus }
     }
@@ -113,10 +106,11 @@ export function cli(options: cli.Options): Adapter.Adapter {
       if (!managedKey) return
 
       const signed = KeyAuthorization.fromRpc(z.encode(CliAuth.keyAuthorization, keyAuthorization))
-      AccessKey.savePending({
+      AccessKeyStore.upsertAuthorization({
         address,
         keyAuthorization: signed,
         privateKey: managedKey.key,
+        state: 'signed',
         store,
       })
 
@@ -151,21 +145,28 @@ export function cli(options: cli.Options): Adapter.Adapter {
 
     async function prepareManagedTransaction(
       client: ReturnType<typeof getClient>,
-      parameters: AccessKey.selectForTransaction.PrepareParameters,
+      parameters: AccessKeyTransaction.create.PrepareParameters,
+      options: {
+        calls?: AccessKeyTransaction.create.Options['calls'] | undefined
+        chainId?: number | undefined
+      } = {},
     ) {
-      const rootAddress = store.getState().accounts[store.getState().activeAccount]?.address
-      const managedKey = rootAddress ? await loadManagedKey(rootAddress) : undefined
-      const account = getAccount({ signable: true })
-      const attempt = await AccessKey.selectForTransaction({
-        account,
+      const state = store.getState()
+      const address = parameters.from ?? state.accounts[state.activeAccount]?.address
+      if (!address) throw new core_Provider.DisconnectedError({ message: 'No active account.' })
+      const managedKey = await loadManagedKey(address)
+      const transaction = await AccessKeyTransaction.create({
+        address,
+        calls: options.calls,
+        chainId: options.chainId ?? state.chainId,
         client,
         store,
       })
-      if (!attempt)
+      if (!transaction)
         throw new core_Provider.UnauthorizedError({
-          message: `Account "${account.address}" cannot sign with an access key.`,
+          message: `Account "${address}" cannot sign with an access key.`,
         })
-      const prepared = await attempt.prepare(parameters)
+      const prepared = await transaction.prepare(parameters)
       if (
         managedKey &&
         managedKey.keyAuthorizationStatus !== 'authorized' &&
@@ -317,10 +318,17 @@ export function cli(options: cli.Options): Adapter.Adapter {
         async sendTransaction(parameters) {
           const { feePayer, ...rest } = parameters
           const client = getClient(typeof feePayer === 'string' ? { feePayer } : {})
-          const { managedKey, prepared } = await prepareManagedTransaction(client, {
-            ...rest,
-            ...(feePayer ? { feePayer: true } : {}),
-          })
+          const { managedKey, prepared } = await prepareManagedTransaction(
+            client,
+            {
+              ...rest,
+              ...(feePayer ? { feePayer: true } : {}),
+            },
+            {
+              calls: parameters.calls as AccessKeyTransaction.create.Options['calls'],
+              chainId: parameters.chainId,
+            },
+          )
           const signed = await prepared.sign()
           if (managedKey && prepared.keyAuthorization)
             await saveManagedKeyStatus(managedKey, 'pending')
@@ -332,10 +340,17 @@ export function cli(options: cli.Options): Adapter.Adapter {
         async sendTransactionSync(parameters) {
           const { feePayer, ...rest } = parameters
           const client = getClient(typeof feePayer === 'string' ? { feePayer } : {})
-          const { managedKey, prepared } = await prepareManagedTransaction(client, {
-            ...rest,
-            ...(feePayer ? { feePayer: true } : {}),
-          })
+          const { managedKey, prepared } = await prepareManagedTransaction(
+            client,
+            {
+              ...rest,
+              ...(feePayer ? { feePayer: true } : {}),
+            },
+            {
+              calls: parameters.calls as AccessKeyTransaction.create.Options['calls'],
+              chainId: parameters.chainId,
+            },
+          )
           const signed = await prepared.sign()
           if (managedKey && prepared.keyAuthorization)
             await saveManagedKeyStatus(managedKey, 'pending')
@@ -345,7 +360,7 @@ export function cli(options: cli.Options): Adapter.Adapter {
           })
           if (managedKey && prepared.keyAuthorization)
             await saveManagedKeyStatus(managedKey, 'authorized')
-          return result as AccessKey.selectForTransaction.SendSyncReturnType
+          return result as AccessKeyTransaction.create.SendSyncReturnType
         },
         async signPersonalMessage({ address, data }) {
           const account = await loadManagedAccount(address)
@@ -354,10 +369,17 @@ export function cli(options: cli.Options): Adapter.Adapter {
         async signTransaction(parameters) {
           const { feePayer, ...rest } = parameters
           const client = getClient(typeof feePayer === 'string' ? { feePayer } : {})
-          const { managedKey, prepared } = await prepareManagedTransaction(client, {
-            ...rest,
-            ...(feePayer ? { feePayer: true } : {}),
-          })
+          const { managedKey, prepared } = await prepareManagedTransaction(
+            client,
+            {
+              ...rest,
+              ...(feePayer ? { feePayer: true } : {}),
+            },
+            {
+              calls: parameters.calls as AccessKeyTransaction.create.Options['calls'],
+              chainId: parameters.chainId,
+            },
+          )
           const signed = await prepared.sign()
           if (managedKey && prepared.keyAuthorization)
             await saveManagedKeyStatus(managedKey, 'pending')
